accountant/accountant/serializers.py
from rest_framework import serializers
from . import models
from django.contrib.auth.models import User

class TransactionSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Transaction
        fields = '__all__' #['id', 'name', 'description', 'created', 'amount', 'tag', 'beneficiary', 'benefactor']

class AccountSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Account
        fields = '__all__'

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.User
        fields = '__all__'