requirements.txt
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.14.0
djongo==1.3.6
image==1.5.33
pillow==10.2.0
pymongo==3.12.1
six==1.16.0
